Revert "Lesson 2 Girfanov"
