--- OTHERS/src/index.js
@@ -1,3 +1,0 @@
-import app from './js/main.js'
-
-app()
