--- PROJECT/0-init/src/css/style.css
@@ -1,186 +1,0 @@
-body{
-    font-family: 'SF Pro Display', sans-serif;
-}
-header{
-    display: flex;
-    background-color: #2f2a2d;
-    justify-content: space-between;
-    color: #fafafa;
-    padding: 30px 80px;
-}
-button:focus{
-    outline: none;
-}
-.logo{
-
-    text-transform: uppercase;
-    font-weight: bold;
-}
-.btn-cart{
-    background-color: #fafafa;
-    padding: 10px 20px;
-    border: 1px solid transparent;
-    color: #2f2a2d;
-    border-radius: 5px;
-    transition: all ease-in-out .4s;
-    cursor: pointer;
-}
-.btn-cart:hover{
-    background-color: transparent;
-    border-color: #fafafa;
-    color: #fafafa;
-}
-.btn-cart, .logo{
-    align-self: center;
-}
-/*.products {*/
-/*display: flex;*/
-/*justify-content: space-between;*/
-/*flex-wrap: wrap;*/
-/*padding: 40px 80px;*/
-/*}*/
-.products{
-    column-gap: 30px;
-    display: grid;
-    grid-template-columns: repeat(auto-fit, 200px);
-    grid-template-rows: 1fr;
-    padding: 40px 80px;
-    justify-content: space-between;
-}
-p {
-    margin: 0 0 5px 0;
-}
-.product-item{
-    display: flex;
-    flex-direction: column;
-    width: 200px;
-    border-radius: 5px;
-    overflow: hidden;
-    margin: 20px 0;
-}
-img {
-    max-width: 100%;
-    height: auto
-}
-.desc {
-    border: 1px solid #c0c0c040;
-    padding: 15px
-}
-
-.desc h1 {
-    font-size: 1em;
-}
-
-.cart{
-    position: relative;
-}
-.cart-block{
-    box-shadow: 0 0 5px rgba(0, 0, 0, 0.62);
-    border-radius: 5px;
-    box-sizing: border-box;
-    right: 0;
-    top: 130%;
-    position: absolute;
-    background-color: white;
-    padding: 20px;
-    color: black;
-    width: 300px;
-}
-
-.invisible{
-    display: none;
-}
-.cart-block:before{
-    content: '';
-    width: 0;
-    height: 0;
-    position: absolute;
-    top: -10px;
-    right: 35px;
-    border-left: 10px solid transparent;
-    border-right: 10px solid transparent;
-    border-bottom: 10px solid white;
-}
-
-.buy-btn, .del-btn{
-    margin-top: 5px;
-    background-color: #2f2a2d;
-    padding: 5px 15px;
-    border: 1px solid transparent;
-    color: #fafafa;
-    border-radius: 5px;
-    transition: all ease-in-out .4s;
-    cursor: pointer;
-}
-.buy-btn:hover, .del-btn:hover{
-    background-color: #fafafa;
-    color: #2f2a2d;
-    border: 1px solid #2f2a2d;
-}
-.cart-item {
-    display: flex;
-    justify-content: space-between;
-}
-.cart-item:not(:last-child){
-    margin-bottom: 20px;
-}
-.product-bio{
-    display: flex;
-}
-.cart-item img{
-    align-self: flex-start;
-    margin-right: 15px;
-}
-.product-single-price{
-    color: #474747;
-    font-size: 0.5em;
-}
-.product-price{
-    margin-left: 30px;
-}
-.product-desc{
-    max-width: 150px;
-}
-.product-quantity {
-    margin-top: 15px;
-    font-size: 0.75em;
-}
-.right-block{
-    text-align: right;
-}
-.btn-search {
-    background-color: transparent;
-    border: none;
-    color: #fafafa;
-    font-size: 1.2em;
-    position: absolute;
-    bottom: 5px;
-    right: 0;
-}
-.search-form{
-    position: relative;
-    margin-right: 50px;
-    display: inline-block;
-}
-.search-field:focus{
-    outline: none;
-}
-.search-field {
-    box-sizing: border-box;
-    width: 200px;
-    color: #fafafa;
-    padding: 10px;
-    background-color: transparent;
-    border: none;
-    border-bottom: 2px solid #fafafa;
-}
-
-.d-flex {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-}
-
-.d-block {
-    display: block
-}
