+++ PROJECT/1_webpack_init/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "src",
+  "name": "1_webpack_init",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -11,7 +11,7 @@
   "license": "ISC",
   "devDependencies": {
     "css-loader": "^3.4.2",
-    "html-webpack-plugin": "^4.0.1",
+    "html-webpack-plugin": "^3.2.0",
     "mini-css-extract-plugin": "^0.9.0",
     "webpack": "^4.42.0",
     "webpack-cli": "^3.3.11"
