+++ PROJECT/1_webpack_init/src/index.js
